requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asttokens
attrs
certifi
charset-normalizer
click
colorama
comm
dataclasses-json
debugpy
decorator
distro
duckduckgo_search
executing
frozenlist
greenlet
h11
httpcore
httpx
idna
ipykernel
ipython

langgraph
langchain-core
langchain-openai
python-dotenv
streamlit
ipython
jsonpatch
jsonpointer
jupyter_client
jupyter_core
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langsmith
marshmallow
matplotlib-inline
multidict
mypy-extensions
nest-asyncio
nltk
numpy
openai
orjson
packaging
pandas
parso
pillow
platformdirs
primp
prompt_toolkit
psutil
pure_eval
pydantic
pydantic_core
Pygments
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
pywin32; platform_system == "Windows"
regex
requests
six
sniffio
SQLAlchemy
stack-data
tabulate
tenacity
tiktoken
tornado
tqdm
traitlets
typing-inspect
typing_extensions
tzdata
urllib3
wcwidth
yarl
autogen
